refactor: extract platform core / domain pack seam (productization phase 1)
Generic multi-agent machinery (tool registry, agent loop, supervisor
helpers) now lives in src/platform; all CRE-specific assets (tools,
prompts, specialist team, result summarizers) move behind the DomainPack
interface in src/packs/cre. Packs are loaded at bootstrap, so a new
vertical ships as data + tools without touching the core.
