+++ orchestrator/src/langgraph/specialists.ts
@@ -1,152 +1,28 @@
-// src/langgraph/specialists.ts — Shared definition of the DealSense agent team.
-// Mirrors the Neuro SAN network (neurosan/registries/dealsense.hocon) so both
-// orchestrators expose the same six specialists over the same 20 tools.
+// src/langgraph/specialists.ts — Platform-side view of the agent team.
+// The actual team definition (supervisor + specialists + prompts) lives in
+// the active domain pack (src/packs/cre/specialists.ts); this module
+// re-exports it for the LangGraph supervisor and keeps the generic helpers
+// (topology builder, JSON-schema→zod). Mirrors the Neuro SAN network
+// (neurosan/registries/dealsense.hocon) since both are generated from the
+// same pack data.
 import { z, ZodTypeAny } from "zod";
+import { getActivePack, type SpecialistSpec } from "../platform/domainPack.js";
 
-export interface SpecialistDef {
-  name: string;
-  description: string; // what the supervisor sees when deciding to delegate
-  prompt: string;      // the specialist's system prompt
-  tools: string[];     // tool names from the registry
-}
+export type SpecialistDef = SpecialistSpec;
 
-export const SUPERVISOR_NAME = "deal_advisor";
+export { SUPERVISOR_NAME, SUPERVISOR_PROMPT, SPECIALISTS } from "../packs/cre/specialists.js";
 
-export const SUPERVISOR_PROMPT = `You are the Deal Advisor, supervisor of the DealSense multi-agent team for
-commercial real estate private equity analysis.
-
-Your team of specialists (each is a tool you can call with an "inquiry"):
-- property_scout: finds and screens deals (search, listing analysis, comps, ranking)
-- risk_analyst: risk scoring, tenant credit, risk decomposition
-- market_analyst: macro data, metro deep dives, market intel, foot traffic
-- financial_modeler: DCF valuation, multi-asset comparison
-- portfolio_manager: portfolio review, VaR, compliance, pipeline
-- deal_writer: investment memos, LOIs
-
-How you work:
-1. Decompose the user's request into a checklist of every distinct part
-   (e.g. "find deals" + "assess risk" + "macro analysis" + "DCF" +
-   "portfolio fit" + "memo" = 6 parts).
-2. Call the specialists in a sensible order; pass each one the specific
-   question plus concrete context from earlier specialists (multi-hop:
-   scout finds a deal, then risk_analyst scores THAT deal, then
-   financial_modeler models THAT deal — include the property title, URL,
-   price and NOI in the context you pass along).
-3. Do NOT produce your final answer until every part of the checklist has
-   been addressed by the corresponding specialist. If a specialist fails,
-   note the failure and continue with the remaining parts.
-   Mandatory mapping — when the user asks for any of these, you MUST call
-   the specialist, never answer it yourself:
-   - find/source/search deals -> property_scout
-   - risk, tenant credit -> risk_analyst
-   - macro, market conditions, metro analysis -> market_analyst
-   - DCF, IRR, valuation, returns -> financial_modeler
-   - portfolio fit, fund, compliance, concentration, VaR -> portfolio_manager
-   - memo, IC memo, LOI, write-up -> deal_writer
-4. Your final answer must be a COMPREHENSIVE markdown report, not a summary:
-   - One ## section per checklist part, in order.
-   - Include the specialists' concrete numbers verbatim (prices, PE scores,
-     cap rates, IRR, equity multiple, risk verdicts, ratings) — never
-     replace numbers with vague phrases.
-   - End with a ## Recommendation section giving a clear verdict.
-5. If specialists disagree (e.g. great PE score but HIGH risk), surface the
-   tension explicitly rather than papering over it.
-6. Only answer questions about commercial real estate analysis. Politely
-   decline anything else without calling any specialist.
-
-Never fabricate data. If a specialist returns an error or empty result,
-report that honestly and suggest what the user could try instead.`;
-
-export const SPECIALISTS: SpecialistDef[] = [
-  {
-    name: "property_scout",
-    description:
-      "Sources and screens commercial real estate deals: marketplace search, listing URL analysis, comparable sales, filtering/ranking.",
-    prompt: `You are the Property Scout, the deal-sourcing specialist of the DealSense team.
-Use search_properties for discovery, analyze_property_url for specific listings,
-comp_analysis for comparables, filter_and_rank to narrow results.
-Always report PE scores and the source marketplace for every property.
-Return structured findings: title, price, cap rate, NOI, PE score, URL.
-
-Resilient search protocol (follow these steps literally):
-1. First call search_properties with the user's full criteria.
-2. If the results do not contain specific property listings satisfying ALL
-   criteria, you MUST call search_properties a SECOND time with a simpler
-   query string that drops the numeric filters but keeps tenant, property
-   type and location (e.g. "NNN Walgreens Texas for sale"). Do not skip
-   this second call.
-3. From all results gathered, present the top 3-5 closest matches under the
-   heading "Closest matches — criteria relaxed (dropped: <filters>)", each
-   with title, PE score, URL, and a note on which original criteria it does
-   or does not satisfy. Presenting closest matches is strongly preferred
-   over reporting nothing.
-4. Only if BOTH searches return no property listings at all, report that
-   no listings were found.
-Never invent listings or fabricate prices/cap rates not present in the data.`,
-    tools: ["search_properties", "analyze_property_url", "comp_analysis", "filter_and_rank"],
-  },
-  {
-    name: "risk_analyst",
-    description:
-      "Assesses deal and tenant risk: market risk scoring, PE deal scoring, tenant credit analysis, risk factor decomposition.",
-    prompt: `You are the Risk Analyst of the DealSense team — the skeptic.
-Use assess_risk for market/deal risk, score_deals for PE scoring,
-tenant_credit_analysis for tenant credit, risk_decomposition for factor breakdowns.
-Always state your verdict explicitly (LOW / MODERATE / ELEVATED / HIGH)
-with the top 3 driving factors. Flag missing or stale data as a risk.`,
-    tools: ["assess_risk", "score_deals", "tenant_credit_analysis", "risk_decomposition"],
-  },
-  {
-    name: "market_analyst",
-    description:
-      "Analyzes macro and local market conditions: FRED/BLS macro data, metro deep dives, market intelligence, foot-traffic patterns.",
-    prompt: `You are the Market Analyst of the DealSense team.
-Use get_macro_data for rates/CPI/unemployment, market_deep_dive for metro analysis,
-market_intel for sector questions, analyze_traffic_patterns for location traffic.
-Cite the data source and as-of date for every number. Summarize trends
-directionally (improving / stable / deteriorating) and say why.`,
-    tools: ["get_macro_data", "market_deep_dive", "market_intel", "analyze_traffic_patterns"],
-  },
-  {
-    name: "financial_modeler",
-    description: "Builds financial models: DCF valuation with sensitivities and multi-asset comparisons.",
-    prompt: `You are the Financial Modeler of the DealSense team.
-Use run_dcf for discounted cash flow models and multi_asset_compare for
-side-by-side comparisons. Always state all assumptions used (discount rate,
-hold period, exit cap, growth). Show key outputs: IRR, NPV, equity multiple.
-If an input is missing, say which assumption you substituted and why.`,
-    tools: ["run_dcf", "multi_asset_compare"],
-  },
-  {
-    name: "portfolio_manager",
-    description:
-      "Institutional portfolio functions: portfolio review, value-at-risk, compliance/allocation checks, acquisition pipeline.",
-    prompt: `You are the Portfolio Manager of the DealSense team.
-Use portfolio_review, portfolio_var, compliance_check and institutional_pipeline.
-Frame every answer in terms of portfolio impact: diversification,
-concentration, allocation limits.`,
-    tools: ["portfolio_review", "portfolio_var", "compliance_check", "institutional_pipeline"],
-  },
-  {
-    name: "deal_writer",
-    description: "Produces investment documents: investment-committee memos and letters of intent.",
-    prompt: `You are the Deal Writer of the DealSense team.
-Use generate_memo for IC memos and generate_loi for letters of intent.
-Only write from facts supplied in the conversation or returned by your tools.
-Keep memos structured: Executive Summary, Deal Terms, Market, Risks, Recommendation.`,
-    tools: ["generate_memo", "generate_loi"],
-  },
-];
-
-/** Same {nodes, edges} shape the Neuro SAN proxy serves, for the UI panel. */
+/** Same {nodes, edges} shape the Neuro SAN proxy serves, for the UI panel.
+ *  Built from the active pack, so it adapts when the pack changes. */
 export function lgNetworkTopology() {
+  const pack = getActivePack();
   const nodes: Array<{ id: string; type: string; parent?: string }> = [
-    { id: SUPERVISOR_NAME, type: "front_man" },
+    { id: pack.supervisorName, type: "front_man" },
   ];
   const edges: Array<{ from: string; to: string }> = [];
-  for (const s of SPECIALISTS) {
-    nodes.push({ id: s.name, type: "specialist", parent: SUPERVISOR_NAME });
-    edges.push({ from: SUPERVISOR_NAME, to: s.name });
+  for (const s of pack.specialists) {
+    nodes.push({ id: s.name, type: "specialist", parent: pack.supervisorName });
+    edges.push({ from: pack.supervisorName, to: s.name });
     for (const t of s.tools) {
       nodes.push({ id: t, type: "tool", parent: s.name });
       edges.push({ from: s.name, to: t });
